Skype bot, multiple improvements, test for Candidate model

+++ client/app/entity/entity.controller.js
@@ -4,11 +4,12 @@
 
 class EntityController {
 
-  constructor($http, $stateParams) {
+  constructor($http, $stateParams, Modal) {
     this.$http = $http;
     this.entities = [];
     this.title = $stateParams.title;
     this.backendUrl = '/api/entities/' + $stateParams.entity + '/';
+    this.Modal = Modal;
 
     $http.get(this.backendUrl).then(response => {
       this.entities = response.data;
@@ -25,9 +26,11 @@ class EntityController {
   }
 
   deleteEntity(entity) {
-    this.$http.delete(this.backendUrl + entity._id).then(response => {
-      _.remove(this.entities, {_id: entity._id});
-    });
+    this.Modal.confirm.delete(() => {
+      this.$http.delete(this.backendUrl + entity._id).then(response => {
+        _.remove(this.entities, {_id: entity._id});
+      });
+    })(entity.name);
   }
 }
